Flytter statusapi til sykmeldingstatus
