+++ src/main/kotlin/no/nav/syfo/aksessering/api/SykmeldingApi.kt
@@ -12,15 +12,16 @@ import io.ktor.routing.route
 import io.ktor.util.KtorExperimentalAPI
 import java.time.LocalDateTime
 import no.nav.syfo.aksessering.SykmeldingService
-import no.nav.syfo.persistering.StatusEvent
-import no.nav.syfo.persistering.SykmeldingStatusEvent
+import no.nav.syfo.sykmeldingstatus.StatusEvent
+import no.nav.syfo.sykmeldingstatus.SykmeldingStatusEvent
+import no.nav.syfo.sykmeldingstatus.SykmeldingStatusService
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
 val log: Logger = LoggerFactory.getLogger("no.nav.syfo.smregister")
 
 @KtorExperimentalAPI
-fun Route.registerSykmeldingApi(sykmeldingService: SykmeldingService) {
+fun Route.registerSykmeldingApi(sykmeldingService: SykmeldingService, sykmeldingStatusService: SykmeldingStatusService) {
     route("/api/v1") {
 
         get("/sykmeldinger") {
@@ -40,7 +41,7 @@ fun Route.registerSykmeldingApi(sykmeldingService: SykmeldingService) {
             val principal: JWTPrincipal = call.authentication.principal()!!
             val subject = principal.payload.subject
             if (sykmeldingService.erEier(sykmeldingsid, subject)) {
-                sykmeldingService.registrerStatus(SykmeldingStatusEvent(sykmeldingsid, LocalDateTime.now(), StatusEvent.BEKREFTET))
+                sykmeldingStatusService.registrerStatus(SykmeldingStatusEvent(sykmeldingsid, LocalDateTime.now(), StatusEvent.BEKREFTET))
                 call.respond(HttpStatusCode.OK)
             } else {
                 call.respond(HttpStatusCode.NotFound)
