+++ src/main/kotlin/no/nav/syfo/persistering/Sykmelding.kt
@@ -30,13 +30,3 @@ fun Sykmelding.toPGObject() = PGobject().also {
     it.type = "json"
     it.value = objectMapper.writeValueAsString(this)
 }
-
-data class SykmeldingStatusEvent(
-    val id: String,
-    val timestamp: LocalDateTime,
-    val event: StatusEvent
-)
-
-enum class StatusEvent {
-    APEN, AVBRUTT, UTGATT, SENDT, BEKREFTET
-}
