+++ src/main/kotlin/no/nav/syfo/sykmeldingstatus/api/SykmeldingSendApi.kt
@@ -25,4 +25,3 @@ fun Route.registerSykmeldingSendApi(sykmeldingStatusService: SykmeldingStatusSer
         }
     }
 }
-
