+++ src/test/kotlin/no/nav/syfo/aksessering/SykmeldingServiceTest.kt
@@ -2,11 +2,12 @@ package no.nav.syfo.aksessering
 
 import java.time.LocalDateTime
 import kotlin.test.assertFailsWith
-import no.nav.syfo.persistering.StatusEvent
-import no.nav.syfo.persistering.SykmeldingStatusEvent
 import no.nav.syfo.persistering.opprettBehandlingsutfall
 import no.nav.syfo.persistering.opprettSykmeldingsdokument
 import no.nav.syfo.persistering.opprettSykmeldingsopplysninger
+import no.nav.syfo.sykmeldingstatus.StatusEvent
+import no.nav.syfo.sykmeldingstatus.SykmeldingStatusEvent
+import no.nav.syfo.sykmeldingstatus.SykmeldingStatusService
 import no.nav.syfo.testutil.TestDB
 import no.nav.syfo.testutil.dropData
 import no.nav.syfo.testutil.testBehandlingsutfall
@@ -22,6 +23,7 @@ class SykmeldingServiceTest : Spek({
 
     val database = TestDB()
     val sykmeldingService = SykmeldingService(database)
+    val sykmeldingStatusService = SykmeldingStatusService(database)
 
     afterGroup {
         database.stop()
@@ -47,21 +49,21 @@ class SykmeldingServiceTest : Spek({
 
         it("Should get bekreftdato = null when status is OPEN") {
             val savedSykmelding = sykmeldingService.hentSykmeldinger("pasientFnr")[0]
-            sykmeldingService.registrerStatus(SykmeldingStatusEvent("uuid", LocalDateTime.now(), StatusEvent.APEN))
+            sykmeldingStatusService.registrerStatus(SykmeldingStatusEvent("uuid", LocalDateTime.now(), StatusEvent.APEN))
             savedSykmelding.bekreftetDato shouldBe null
         }
 
         it("Should get bekreftetDato") {
             val confirmedDateTime = LocalDateTime.now()
-            sykmeldingService.registrerStatus(SykmeldingStatusEvent("uuid", confirmedDateTime, StatusEvent.BEKREFTET))
+            sykmeldingStatusService.registrerStatus(SykmeldingStatusEvent("uuid", confirmedDateTime, StatusEvent.BEKREFTET))
             val savedSykmelding = sykmeldingService.hentSykmeldinger("pasientFnr")[0]
             savedSykmelding.bekreftetDato shouldEqual confirmedDateTime
         }
 
         it("Should get bekreftetDato when newest status is SENDT") {
             val confirmedDateTime = LocalDateTime.now().minusDays(1)
-            sykmeldingService.registrerStatus(SykmeldingStatusEvent("uuid", confirmedDateTime, StatusEvent.BEKREFTET))
-            sykmeldingService.registrerStatus(SykmeldingStatusEvent("uuid", confirmedDateTime.plusDays(1), StatusEvent.SENDT))
+            sykmeldingStatusService.registrerStatus(SykmeldingStatusEvent("uuid", confirmedDateTime, StatusEvent.BEKREFTET))
+            sykmeldingStatusService.registrerStatus(SykmeldingStatusEvent("uuid", confirmedDateTime.plusDays(1), StatusEvent.SENDT))
             val savedSykmelding = sykmeldingService.hentSykmeldinger("pasientFnr")[0]
             savedSykmelding.bekreftetDato shouldEqual confirmedDateTime
         }
@@ -69,8 +71,8 @@ class SykmeldingServiceTest : Spek({
         it("Should throw error when inserting same status") {
             val confirmedDateTime = LocalDateTime.now()
             val status = SykmeldingStatusEvent("uuid", confirmedDateTime, StatusEvent.BEKREFTET)
-            sykmeldingService.registrerStatus(status)
-            assertFailsWith(PSQLException::class) { sykmeldingService.registrerStatus(status) }
+            sykmeldingStatusService.registrerStatus(status)
+            assertFailsWith(PSQLException::class) { sykmeldingStatusService.registrerStatus(status) }
         }
     }
 })
